LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-blank-navigation-expected.txt



FAIL entries() after navigation to about:blank assert_equals: expected 2 but got 0
